pyproject.toml
[project]
name = "sharp"
version = "4.0.0"
description = "Scalable Heterogeneous Architecture for Reproducible Performance"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE.md"}
authors = [
    {name = "Eitan Frachtenberg", email = "eitan.frachtenberg@hpe.com"},
    {name = "Viyom Mittal", email = "viyom.mittal@hpe.com"}
]

dependencies = [
    "pyyaml>=6.0.0",
    "numpy>=2.0.0",
    "scipy>=1.15.0",
    "scikit-learn>=1.5.0",
    "isodate>=0.6.0",
    "arviz>=0.21.0",
    "pydantic>=2.0.0",
    "ruptures>=1.1.0",
    "statsmodels>=0.14.0",
    "polars>=1.0.0",
    "shiny>=1.0.0",
    "shinyswatch>=0.7.0",
    "plotly>=5.0.0",
    "graphviz>=0.20.0",
    "pillow>=12.0.0",
]

[project.scripts]
sharp = "src.cli.sharp:main"
launch = "src.cli.launch:main"
build = "src.cli.build:main"
compare = "src.cli.compare:main"
gui = "src.gui.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
managed = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, multiple components)",
    "smoke: Smoke tests (minimal end-to-end validation)",
    "slow: Slow tests (profiling, large datasets)",
    "performance: Performance regression tests",
]
filterwarnings = [
    "ignore::RuntimeWarning:numpy",
    "ignore::RuntimeWarning:statsmodels",
    "ignore::RuntimeWarning:scipy",
]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "flask>=2.2.0",
    "werkzeug>=3.0.0",
    "numba>=0.61.0",
    "mpi4py>=4.0.0",
    "types-pyyaml>=6.0.12.20250915",
]
